# documentation/openapi.yaml
openapi: '3.0.3'
info:
  title: OpenEvents
  description: An Express API for OpenEvents
  contact:
    email: danimr1999@gmail.com
    name: Daniel Muelle Ramos
  license:
    name: MIT
    url: https://www.mit.edu/~amini/LICENSE.md
  version: 1.0.0
servers:
- url: http://localhost:3000/
  description: Development
paths:
...
  /users/search:
    get:
      tags:
        - Users
      summary: Searches users with a name, last name or email matching the value of the query parameter.
      security:
          - BearerAuth: []
      parameters:
          - in: query
            name: text
            schema:
              type: string
            required: true
            description: Text to search on the name, last name or email fields of a user to match.
      responses:
        "200":
          description: List of users that match with the text on their name, last name or email field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListUserWithId'
        "400":
          description: No text introduced.
  /events:
    get:
      tags:
        - Events
      summary: Gets all future events.
      description: Gets all the events whose start date is greater than now. 
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of future events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListEventWithId'
    post:
      tags:
        - Events
      summary: Creates an event.
      description: Depending on some parameters, there are fields that can be optional. 
        If an event is online, the link is required and the location and the max_attendees are optional.
        If an event is face-to-face, the link is optional and the location and the max_attendees are required.
      security:
        - BearerAuth: []
      requestBody:
        description: Event information in JSON format.
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/RequestEvent'
      responses:
        "201":
          description: Created event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEventWithId'
  /events/search:
    get:
      tags:
        - Events
      summary: Searches events with a title and/or location matching the value of the queries parameters.
      security:
          - BearerAuth: []
      parameters:
          - in: query
            name: title
            schema:
              type: string
            required: false
            description: Title to search.
          - in: query
            name: location
            schema:
              type: string
            required: false
            description: Location to search.
      responses:
        "200":
          description: List of events that match with the title and/or location queried.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListEventWithId'
        "400":
          description: Invalid title and/or location.
  /events/{event_id}:
      get:
        tags:
          - Events
        summary: Gets an event by ID.
        security:
          - BearerAuth: []
        parameters:
          - in: path
            name: event_id
            schema:
              type: integer
            required: true
            description: ID of the event to get
        responses:
          "200":
            description: Event matching with the specified ID.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ResponseEventWithId'
          "400":
            description: Event does not exist or was not found.
      put:
        tags:
          - Events
        summary: Edits specified fields of an event if the authenticated user is the owner of the event.
        security:
          - BearerAuth: []
        parameters:
          - in: path
            name: event_id
            schema:
              type: integer
            required: true
            description: ID of the event to edit 
        requestBody:
          description: Event information in JSON format. Only the fields to be updated are required.
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestEvent'
        responses:
          "200":
            description: Event with the updated information.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ResponseEventWithId'
          "400":
            description: Any of the specified event fields to modify are not valid.
          "403":
            description: Authenticated user trying to modify the specified event is not the owner of it.
          "404":
            description: Specified event does not exist or was not found.
      delete:
        tags:
          - Events
        summary: Deletes an event with matching ID.
        description: Deletes an event by ID if the authenticated user is the owner of it.
        security:
          - BearerAuth: []
        parameters:
          - in: path
            name: event_id
            schema:
              type: integer
            required: true
            description: ID of the event to delete
        responses:
          "200":
            description: Event deleted successfully.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      description: Message notifying that the action has been completed successfully.
                      example: "Event was deleted successfully"
          "403":
            description: Authenticated user is not the owner of the specified event.
          "404":
            description: Event to be deleted does not exist or was not found.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ResponseListUserWithId:
      type: array
      items:
        type: object
        properties:
            id:
              type: number
              example: 1
            name:
              type: string
              example: "Jane"
            last_name:
              type: string
              example: "Doe"
            email:
              type: string
              example: "janedoe@openevents.com"
            image_url:
              type: string
              example: "https://i.imgur.com/ghy8Xx1.png"
    RequestEvent:
      type: object
      properties:
          title:
            type: string
            example: "Drawing 101: Portraits"
          image_url:
            type: string
            example: "https://i.imgur.com/JprpLyc.jpg"
          format:
            type: number
            example: 1
          link:
            type: string
            example: "https://www.zoom.com/"
          location:
            type: string
            example: "Barcelona"
          description:
            type: string
            example: "Learn the basics of portrait drawing with a professional artist 
              with +10 years of experience"
          start_date:
            type: string
            example: "2022-07-23 10:00"
          end_date:
            type: string
            example: "2022-07-25 10:00"
          max_attendees:
            type: number
            example: 120
          ticket_price:
            type: number
            example: 4.99
          category: 
            type: string
            example: 'Art'
    ResponseEventWithId:
      type: object
      properties:
          id:
            type: number
            example: 12
          title:
            type: string
            example: "Drawing 101: Portraits"
          owner_id:
            type: number
            example: 1
          creation_date:
            type: string
            example: "2022-07-21T23:26:19.000Z"
          image_url:
            type: string
            example: "https://i.imgur.com/JprpLyc.jpg"
          format:
            type: number
            example: 1
          link:
            type: string
            example: "https://www.zoom.com/"
          location:
            type: string
            example: "Barcelona"
          description:
            type: string
            example: "Learn the basics of portrait drawing with a professional artist 
              with +10 years of experience"
          start_date:
            type: string
            example: "2022-07-23T08:00:00.000Z"
          end_date:
            type: string
            example: "2022-07-25T08:00:00.000Z"
          max_attendees:
            type: number
            example: 120
          ticket_price:
            type: number
            example: 4.99
          category: 
            type: string
            example: 'Art'
    ResponseListEventWithId:
      type: array
      items:
        type: object
        properties:
          id:
            type: number
            example: 12
          title:
            type: string
            example: "Drawing 101: Portraits"
          owner_id:
            type: number
            example: 1
          creation_date:
            type: string
            example: "2022-07-21T23:26:19.000Z"
          image_url:
            type: string
            example: "https://i.imgur.com/JprpLyc.jpg"
          format:
            type: number
            example: 1
          link:
            type: string
            example: "https://www.zoom.com/"
          location:
            type: string
            example: "Barcelona"
          description:
            type: string
            example: "Learn the basics of portrait drawing with a professional artist 
              with +10 years of experience"
          start_date:
            type: string
            example: "2022-07-23T08:00:00.000Z"
          end_date:
            type: string
            example: "2022-07-25T08:00:00.000Z"
          max_attendees:
            type: number
            example: 120
          ticket_price:
            type: number
            example: 4.99
          category: 
            type: string
            example: 'Art'
...